Rename C program

+++ Study/C/2_Library_Math/geron.c
@@ -5,12 +5,12 @@
 int main()
 {
     int a, b, c;
-    float p, s;
+    double p, s;
     printf("Vvedi 3 storoni: ");
     scanf("%d%d%d", &a, &b, &c);
-    p = (a + b + c) / 3;
+    p = (a + b + c) / 3.0;
     s = sqrt(p * (p - a) * (p - b) * (p - c));
-    printf("\n S = %.5f P = %.2f", s, p * 2);
+    printf("\n S = %.5lf P = %.2lf", s, p * 2);
     getch();
     return 0;
 }
